Moved the client into its own lib.
